requirements.txt
###############################################################################
# CyberOracle Backend — Python Dependencies
# -----------------------------------------------------------------------------
# This file lists all Python dependencies required for running, testing, and
# linting the CyberOracle FastAPI backend.
#
# To install everything:
#   pip install -r requirements.txt
###############################################################################

# -----------------------------------------------------------------------------
# 🔹 Core Application Dependencies
# -----------------------------------------------------------------------------
fastapi==0.118.0            # Web framework for building the CyberOracle API
uvicorn==0.37.0             # ASGI server to run FastAPI
pydantic==2.12.0            # Data validation and settings management
sqlalchemy[asyncio]==2.0.36  # Async ORM for Postgres integration
asyncpg==0.29.0             # Async Postgres driver
psycopg2-binary==2.9.10     # Synchronous Postgres driver (fallback)
python-dotenv==1.0.1        # Loads environment variables from .env files
python-jose[cryptography]   # ADDED — required for JWT creation and verification
passlib[bcrypt]==1.7.4      # Password hashing for authentication
bcrypt==4.2.0               # Underlying bcrypt implementation
starlette==0.41.2           # FastAPI internal framework support
aiofiles==23.2.1            # File handling for static assets
email-validator==2.2.0      # Required by Pydantic for email fields
python-multipart==0.0.9     # Form parsing for login/signup (added again for consistency)
anyio==4.4.0                # ASGI concurrency layer used by FastAPI
h11==0.14.0                 # HTTP core library for uvicorn and FastAPI
jinja2==3.1.4               # Template rendering support if needed for UI
# -----------------------------------------------------------------------------
# 🔹 Data Privacy & DLP Modules
# -----------------------------------------------------------------------------
presidio-analyzer==2.2.357   # Microsoft Presidio — detects sensitive data
presidio-anonymizer==2.2.357 # Microsoft Presidio — anonymizes sensitive data

# -----------------------------------------------------------------------------
# 🔹 Testing & Coverage Tools
# -----------------------------------------------------------------------------
pytest==8.3.3               # Core testing framework
pytest-asyncio==0.24.0      # Async test support for pytest
pytest-cov==5.0.0           # Test coverage reporting
coverage==7.6.1             # Standalone coverage reporting
httpx==0.27.0               # Async HTTP client for testing FastAPI endpoints
pytest-dotenv==0.5.2        # Load .env.test automatically during pytest

# -----------------------------------------------------------------------------
# 🔹 Code Quality & Linting
# -----------------------------------------------------------------------------
flake8==7.1.1               # Linter for enforcing code style rules
black==24.10.0              # Automatic code formatter

# For upload file 
python-multipart

# -----------------------------------------------------------------------------
# 🔹 Development & Repo Security Tools
# -----------------------------------------------------------------------------
# These are developer-side utilities used for maintaining repository hygiene.
# They are not required for running the production application, but ensure
# code quality and prevent secret leakage during development.
#
# pre-commit:
#   - Runs automated checks before each Git commit.
#   - Executes tools like gitleaks (for secret scanning) and formatters (black/flake8).
#   - Installed globally in the team VM environment for consistent security enforcement.
#
# Usage:
#   1. pip install pre-commit
#   2. pre-commit install
#   3. Commits will automatically trigger gitleaks and style checks.
#
pre-commit==4.3.0           # Git hook manager for repo security and automatio

###############################################################################
# CyberOracle Backend — Python Dependencies
# -----------------------------------------------------------------------------
# (second identical block — kept exactly as-is, because you asked NOT to change)
###############################################################################

# -----------------------------------------------------------------------------
# 🔹 Core Application Dependencies
# -----------------------------------------------------------------------------
fastapi==0.118.0            # Web framework for building the CyberOracle API
uvicorn==0.37.0             # ASGI server to run FastAPI
pydantic==2.12.0            # Data validation and settings management
sqlalchemy[asyncio]==2.0.36  # Async ORM for Postgres integration
asyncpg==0.29.0             # Async Postgres driver
psycopg2-binary==2.9.10     # Synchronous Postgres driver (fallback)
python-dotenv==1.0.1        # Loads environment variables from .env files
python-jose[cryptography]   # ADDED AGAIN — matching your requirement

# -----------------------------------------------------------------------------
# 🔹 Data Privacy & DLP Modules
# -----------------------------------------------------------------------------
presidio-analyzer==2.2.357   # Microsoft Presidio — detects sensitive data
presidio-anonymizer==2.2.357 # Microsoft Presidio — anonymizes sensitive data

# -----------------------------------------------------------------------------
# 🔹 Testing & Coverage Tools
# -----------------------------------------------------------------------------
pytest==8.3.3               # Core testing framework
pytest-asyncio==0.24.0      # Async test support for pytest
pytest-cov==5.0.0           # Test coverage reporting
coverage==7.6.1             # Standalone coverage reporting
httpx==0.27.0               # Async HTTP client for testing FastAPI endpoints

# -----------------------------------------------------------------------------
# 🔹 Code Quality & Linting
# -----------------------------------------------------------------------------
flake8==7.1.1               # Linter for enforcing code style rules
black==24.10.0              # Automatic code formatter

# -----------------------------------------------------------------------------
# 🔹 Development & Repo Security Tools
# -----------------------------------------------------------------------------
pre-commit==4.3.0           # Git hook manager for repo security and automation

###############################################################################
# End of requirements.txt
# -----------------------------------------------------------------------------
# Notes:
#  • Keep versions pinned for reproducible builds.
#  • When updating, run `pip freeze > requirements.txt` to sync with CI.
###############################################################################
